Add support for API access using Service Account
Previously, it was required to configure an OAuth 2.0 client using the authorization code flow and obtain a refresh token beforehand.
Additionally, the refresh token has the potential to become invalidated (e.g., if not used for 6 months).
Therefore, as an alternative approach, I have added support for Service Account, which is suitable for machine-to-machine API access.